start postgres as a child of process
We currently use pg_ctl which starts postgres in the background. This
means if we shutdown too fast the process is left lieing around. By
directly starting the postgres binary it should get cleaned up.

To correctly support this logic we had to implement a part of what
pg_ctl start does (parsing of the postmaster PID file).n

Test Plan: go test